feat(reset-state): add command to ag-solo to reset SwingSet
Note that the ag-cosmos-helper state is not reset.
